Rename the SenatorSelectInput to SenatorSelector
